refactor: consolidate ProfilePage, ProfileInfo, and ProfileEditForm
