+++ roamance/src/app/profile/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ProfilePage } from '@/components/profile/profile-page';
+import { ProfilePage } from '@/components/profile';
 
 export default function Profile() {
   return <ProfilePage />;
